Import catogram changes
Credits to Catogram contributors

+++ TMessagesProj/src/main/java/org/telegram/ui/Charts/BaseChartView.java
@@ -346,7 +346,7 @@ public abstract class BaseChartView<T extends ChartData, L extends LineViewData>
         chartArea.set(chartStart - HORIZONTAL_PADDING, 0, chartEnd + HORIZONTAL_PADDING, getMeasuredHeight() - chartBottom);
 
         if (chartData != null) {
-            bottomSignatureOffset = (int) (AndroidUtilities.dp(20) / ((float) pickerWidth / chartData.x.length));
+            bottomSignatureOffset = (int) (AndroidUtilities.dp(20) / (pickerWidth / chartData.x.length));
         }
         measureHeightThreshold();
     }
@@ -354,7 +354,7 @@ public abstract class BaseChartView<T extends ChartData, L extends LineViewData>
     private void measureHeightThreshold() {
         int chartHeight = getMeasuredHeight() - chartBottom;
         if (animateToMaxHeight == 0 || chartHeight == 0) return;
-        thresholdMaxHeight = ((float) animateToMaxHeight / chartHeight) * SIGNATURE_TEXT_SIZE;
+        thresholdMaxHeight = (animateToMaxHeight / chartHeight) * SIGNATURE_TEXT_SIZE;
     }
 
 
@@ -553,7 +553,7 @@ public abstract class BaseChartView<T extends ChartData, L extends LineViewData>
 
         float additionalOutAlpha = 1f;
         if (n > 2) {
-            float v = (a.values[1] - a.values[0]) / (float) (currentMaxHeight - currentMinHeight);
+            float v = (a.values[1] - a.values[0]) / (currentMaxHeight - currentMinHeight);
             if (v < 0.1) {
                 additionalOutAlpha = v / 0.1f;
             }
@@ -581,7 +581,7 @@ public abstract class BaseChartView<T extends ChartData, L extends LineViewData>
 
         float additionalOutAlpha = 1f;
         if (n > 2) {
-            float v = (a.values[1] - a.values[0]) / (float) (currentMaxHeight - currentMinHeight);
+            float v = (a.values[1] - a.values[0]) / (currentMaxHeight - currentMinHeight);
             if (v < 0.1) {
                 additionalOutAlpha = v / 0.1f;
             }
